# backend/api-gateway/openapi.yaml
openapi: 3.0.3
info:
  title: API Gateway
  description: |
    API Gateway
  version: v1
servers:
  - url: "http://localhost:8000"
    description: localhost
tags:
  - name: Todo
    description: Todo API
paths:
  /todos:
    get:
      summary: Browse todo items
      description: Browse todo items
      tags:
        - Todo
      parameters:
        - $ref: "#/components/parameters/offset"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          $ref: "#/components/responses/Todos"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
    post:
      summary: Create a new todo item
      description: Create a new todo item
      tags:
        - Todo
      requestBody:
        $ref: "#/components/requestBodies/NewTodo"
      responses:
        "200":
          description: Ok
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
components:
  parameters:
    offset:
      name: offset
      in: query
      description: Pagination offset
      schema:
        type: integer
      example: 0
    limit:
      name: limit
      in: query
      description: Pagination limit
      schema:
        type: integer
      example: 10
  requestBodies:
    NewTodo:
      description: New Todo
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/NewTodo"
  responses:
    Todos:
      description: Todos
      content:
        application/json:
          schema:
            type: object
            properties:
              meta:
                type: object
                description: Metadata
                properties:
                  offset:
                    type: integer
                    description: Pagination offset
                    example: 0
                  limit:
                    type: integer
                    description: Pagination limit
                    example: 10
              data:
                type: array
                description: Data
                items:
                  $ref: "#/components/schemas/Todo"
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Todo:
      type: object
      description: Todo
      required:
        - id
        - name
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          description: Todo ID
          example: 1
        name:
          type: string
          description: Todo name
          example: Demo
        description:
          type: string
          description: Todo description
          example: Prepare the demo
        created_at:
          type: string
          description: Creation timestamp
          example: "2021-05-01T17:30:00Z"
        updated_at:
          type: string
          description: Update timestamp
          example: "2021-05-01T18:00:00Z"
    NewTodo:
      type: object
      description: New Todo
      required:
        - name
      properties:
        name:
          type: string
          description: Todo name
          example: Demo
        description:
          type: string
          description: Todo description
          example: Prepare the demo
    Error:
      type: object
      description: Error response
      required:
        - error
      properties:
        error:
          type: object
          description: Error payload
          required:
            - code
          properties:
            code:
              type: string
              description: Error code
              example: <error_code>
            details:
              type: object
              description: Error details